integrate TownPortalScroll into object plugin
